put tensor storage back
