+++ test/test_tensor_storage.cpp
@@ -1,5 +1,5 @@
 #include <utest/utest.h>
-#include <nano/tensor/storage2.h>
+#include <nano/tensor/storage.h>
 
 using namespace nano;
 
@@ -18,14 +18,14 @@ static void storage_must_match(const tlhs& lhs, const trhs& rhs)
     UTEST_CHECK_EIGEN_CLOSE(map_lhs, map_rhs, 1e-12);
 }
 
-static vector_t data0 = vector_t::Constant(7, 0);
-static const vector_t data1 = vector_t::Constant(7, 1);
-static const vector_t data2 = vector_t::Constant(7, 2);
-
 UTEST_BEGIN_MODULE(test_tensor_storage)
 
 UTEST_CASE(vector_init)
 {
+    vector_t data0 = vector_t::Constant(7, 0);
+    const vector_t data1 = vector_t::Constant(7, 1);
+    const vector_t data2 = vector_t::Constant(7, 2);
+
     // vector(size)
     {
         const auto vector = vector_storage_t{5};
@@ -75,6 +75,10 @@ UTEST_CASE(vector_init)
 
 UTEST_CASE(carray_init)
 {
+    vector_t data0 = vector_t::Constant(7, 0);
+    const vector_t data1 = vector_t::Constant(7, 1);
+    const vector_t data2 = vector_t::Constant(7, 2);
+
     // carray(carray)
     {
         const auto carray1 = carray_storage_t{data0.data(), data0.size()};
@@ -106,6 +110,10 @@ UTEST_CASE(carray_init)
 
 UTEST_CASE(marray_init)
 {
+    vector_t data0 = vector_t::Constant(7, 0);
+    const vector_t data1 = vector_t::Constant(7, 1);
+    const vector_t data2 = vector_t::Constant(7, 2);
+
     // marray(marray)
     {
         const auto marray1 = marray_storage_t{data0.data(), data0.size()};
@@ -128,6 +136,10 @@ UTEST_CASE(marray_init)
 
 UTEST_CASE(vector_copy)
 {
+    vector_t data0 = vector_t::Constant(7, 0);
+    const vector_t data1 = vector_t::Constant(7, 1);
+    const vector_t data2 = vector_t::Constant(7, 2);
+
     auto vector = vector_storage_t{data2};
     storage_must_match(vector, data2);
     UTEST_CHECK_NOT_EQUAL(vector.data(), data2.data());
@@ -153,6 +165,10 @@ UTEST_CASE(vector_copy)
 
 UTEST_CASE(marray_copy)
 {
+    vector_t data0 = vector_t::Constant(7, 0);
+    const vector_t data1 = vector_t::Constant(7, 1);
+    const vector_t data2 = vector_t::Constant(7, 2);
+
     const auto vector = vector_storage_t{data2};
     const auto marray = marray_storage_t{data0.data(), data0.size()};
     const auto carray = carray_storage_t{data1.data(), data1.size()};
@@ -186,6 +202,10 @@ UTEST_CASE(marray_copy)
 
 UTEST_CASE(resize)
 {
+    vector_t data0 = vector_t::Constant(7, 0);
+    const vector_t data1 = vector_t::Constant(7, 1);
+    const vector_t data2 = vector_t::Constant(7, 2);
+
     auto vector = vector_storage_t{data2};
     auto marray = marray_storage_t{data0.data(), data0.size()};
     auto carray = carray_storage_t{data1.data(), data1.size()};
